feat: Enable Pulumi to be run as an execute via Bazel
This commit shows how to execute a binary from a nested NodeJS project
for Pulumi.

Pulumi can be executed as a binary in addition to being executed via the
CLI. Executing Pulumi as a binary process allows us to better integrate
Pulumi with our code, and bring our infrastructure closer to our code.

Future work has to resolve issues such as passing in of image tags for
deployments and so forth.