feat: Tie everything together for releasing
This adds in `.swcrc` file generation,
passing along of version tags to Pulumi,
execution of Pulumi via the automation API,
and more.

We also added basic stamping of the workspace to help
with versioning and rebuilding as needed.

Our Pulumi is switched to using the Pulumi server,
and not store state in S3. This is due to issues arrising
from failed creation events, which are treated better with
the Pulumi service backend. We also add back in the Pulumi
yaml files, to allow for manual interaction of the Pulumi
CLI with the Pulumi service, for our stack.

We fix the image generation logic to properly support generation
of multi-arch images.